requirements.txt
# Standard libraries
python-dotenv

# Redis
redis

# Celery
celery
